+++ app/javascript/flavours/glitch/actions/push_notifications/setter.js
@@ -1,9 +1,7 @@
-import axios from 'axios';
-
 export const SET_BROWSER_SUPPORT = 'PUSH_NOTIFICATIONS_SET_BROWSER_SUPPORT';
 export const SET_SUBSCRIPTION = 'PUSH_NOTIFICATIONS_SET_SUBSCRIPTION';
 export const CLEAR_SUBSCRIPTION = 'PUSH_NOTIFICATIONS_CLEAR_SUBSCRIPTION';
-export const ALERTS_CHANGE = 'PUSH_NOTIFICATIONS_ALERTS_CHANGE';
+export const SET_ALERTS = 'PUSH_NOTIFICATIONS_SET_ALERTS';
 
 export function setBrowserSupport (value) {
   return {
@@ -25,28 +23,12 @@ export function clearSubscription () {
   };
 }
 
-export function changeAlerts(key, value) {
+export function setAlerts (key, value) {
   return dispatch => {
     dispatch({
-      type: ALERTS_CHANGE,
+      type: SET_ALERTS,
       key,
       value,
-    });
-
-    dispatch(saveSettings());
-  };
-}
-
-export function saveSettings() {
-  return (_, getState) => {
-    const state = getState().get('push_notifications');
-    const subscription = state.get('subscription');
-    const alerts = state.get('alerts');
-
-    axios.put(`/api/web/push_subscriptions/${subscription.get('id')}`, {
-      data: {
-        alerts,
-      },
     });
   };
 }
